+++ adminer/drivers/mysql.inc.php
@@ -58,6 +58,10 @@ if (!defined('Adminer\DRIVER')) {
 			function inTransaction(): bool {
 				return false; // MySQLi exposes no API for it and this class can't extend SqlDb to inherit the default
 			}
+
+			function begin(): bool {
+				return $this->begin_transaction(); // commit() and rollback() are inherited from MySQLi
+			}
 		}
 
 	} elseif (extension_loaded("mysql") && !((ini_bool("sql.safe_mode") || ini_bool("mysql.allow_local_infile")) && extension_loaded("pdo_mysql"))) {